refactor!: Un-export `Timer` (#90)
This PR removes the internal `Timer` class from the exports. It is a
very simple wrapper over the `Stopwatch` API and does not provide any
additional functionality. But it does causes conflicts when try to use
Flame's `Timer` or dart:async's `Timer`. By not exporting it, we can
save the users from these conflicts.